+++ pumlgenerator/src/main/kotlin/UMLGenerator.kt
@@ -1,10 +1,13 @@
+import com.google.devtools.ksp.findActualType
 import com.google.devtools.ksp.processing.KSPLogger
 import com.google.devtools.ksp.symbol.KSClassDeclaration
 import com.google.devtools.ksp.symbol.KSFile
 import com.google.devtools.ksp.symbol.KSFunctionDeclaration
 import com.google.devtools.ksp.symbol.KSPropertyDeclaration
+import com.google.devtools.ksp.symbol.KSTypeAlias
 import com.google.devtools.ksp.symbol.KSVisitorVoid
 import com.google.devtools.ksp.validate
+import uml.fullQualifiedName
 
 class UMLGenerator(val logger: KSPLogger, val diagrams: ClassDiagramDescription, val options: Options) : KSVisitorVoid() {
     override fun visitFile(file: KSFile, data: Unit) {
@@ -14,7 +17,7 @@ class UMLGenerator(val logger: KSPLogger, val diagrams: ClassDiagramDescription,
         }
 
         logger.v { "${file.fileName} -> ${file.declarations.joinToString()}" }
-        val ignoredDeclarations = file.declarations.filter { it !is KSClassDeclaration && it !is KSFunctionDeclaration && it !is KSPropertyDeclaration }.toList()
+        val ignoredDeclarations = file.declarations.filter { it !is KSClassDeclaration && it !is KSFunctionDeclaration && it !is KSPropertyDeclaration && it !is KSTypeAlias }.toList()
         if (ignoredDeclarations.isNotEmpty()) {
             logger.w { "The following declarations were ignored since only class declarations are considered: ${ignoredDeclarations.joinToString()}" }
         }
@@ -54,6 +57,20 @@ class UMLGenerator(val logger: KSPLogger, val diagrams: ClassDiagramDescription,
         logger.v { "VisitClassDeclaration of: ${classDeclaration.qualifiedName?.asString()}" }
 
         diagrams.addClass(classDeclaration = classDeclaration)
+
+    }
+
+    override fun visitTypeAlias(typeAlias: KSTypeAlias, data: Unit) {
+        if (!typeAlias.validate()) {
+            logger.i { "visitTypeAlias(): $typeAlias is not valid" }
+            return
+        }
+        if (!options.isValid(typeAlias, logger)) {
+            return
+        }
+        logger.v { "visitTypeAlias: ${typeAlias.fullQualifiedName} (alias of ${typeAlias.findActualType().fullQualifiedName})" }
+
+        diagrams.addTypeAlias(typeAlias = typeAlias)
 
     }